requirements.txt
python-dotenv
fastapi
uvicorn
pymilvus
pymilvus[model]
mistralai
pypdf
langchain
langchain_community 
langchain_huggingface 
langchain-milvus
langchain_mistralai
torch
datasets
peft
nltk
python-multipart
cryptography